refactor: mylibrary 관련 파일 구조 개선
