+++ src/app/(with-header)/my-library/(detail)/[id]/layout.tsx
